after loading cl-change-case in and empty sbcl it works
[martin@fedora cl-cpp-generator2]$ sbcl
This is SBCL 2.1.0.54-252811a68, an implementation of ANSI Common Lisp.
More information about SBCL is available at <http://www.sbcl.org/>.

SBCL is free software, provided as is, with absolutely no warranty.
It is mostly in the public domain; some portions are provided under
BSD-style licenses.  See the CREDITS and COPYING files in the
distribution for more information.
* (ql:quickload "cl-change-case")
To load "cl-change-case":
  Load 1 ASDF system:
    cl-change-case
; Loading "cl-change-case"
..................................................
[package flexi-streams]...........................
..................................................
..................................................
..................................................
..................................................
[package cl-ppcre-unicode]........................
[package cl-change-case].
("cl-change-case")
*