too bad:
you cannot overload a method or operator in C++ specifically for when a bit field within a union or struct is written to, such as status.reserved = <value>. C++ does not provide a direct mechanism to intercept or overload assignments to individual members of a struct or union, including bit fields. The language simply does not support this level of granularity in operator overloading or method overriding.